+++ src/django/api/models.py
@@ -349,11 +349,14 @@ class FacilityList(models.Model):
     updated_at = models.DateTimeField(auto_now=True)
 
     def __str__(self):
-        if self.source.contributor is None:
-            return '{0} ({1})'.format(self.name, self.id)
+        try:
+            if self.source.contributor is None:
+                return '{0} ({1})'.format(self.name, self.id)
 
-        return '{0} - {1} ({2})'.format(
-            self.source.contributor.name, self.name, self.id)
+            return '{0} - {1} ({2})'.format(
+                self.source.contributor.name, self.name, self.id)
+        except Source.DoesNotExist:
+            return '{0} [NO SOURCE] ({1})'.format(self.name, self.id)
 
 
 class FacilityListItem(models.Model):